% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bergm-package.R
\docType{package}
\name{Bergm-package}
\alias{Bergm-package}
\alias{Bergm}
\title{Bayesian exponential random graph models}
\description{
\code{Bergm} provides a range of tools to analyse 
Bayesian exponential random graph models using advanced 
computational methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://acaimo.github.io/Bergm/}
}

}
