\name{bergmS.output}
\alias{bergmS.output}
\title{
MCMC output diagnostics for the Bayesian model selection algorithm
}
\description{
This function returns posterior model and parameter density estimates and creates simple diagnostic plots for the MCMC produced by the function \code{bergmS}. 
}
\usage{
bergmS.output(x, 
              ...)
}
\arguments{
  \item{x}{ 
  an \code{R} object of class \code{bergmS}.
}
  \item{...}{
additional arguments, to be passed to lower-level functions.
}
}
\seealso{
  \code{\link{bergmS}}.
}

