\name{Bergm-package}
\Rdversion{2.4}
\alias{Bergm-package}
\docType{package}
\title{
Bayesian analysis for exponential random graph models
}
\description{
\code{Bergm} is a collection of functions for Bayesian exponential random graph models.
}
\author{
Alberto Caimo <acaimo.stats@gmail.com> and Nial Friel <nial.friel@ucd.ie>.
}

\references{
Caimo, A. and Friel, N. (2012b), ``Bergm: Bayesian Exponential Random Graphs in R,'' \emph{Tech. rep., University College Dublin}. Available in e-print format at http://arxiv.org/abs/1201.2770.

Caimo, A. and Friel, N. (2012a), ``Bayesian model selection for exponential random graph models,'' \emph{Tech. rep., University College Dublin}. Available in e-print format at http://arxiv.org/abs/1201.2337.

Caimo, A. and Friel, N. (2011), ``Bayesian inference for exponential random graph models,'' \emph{Social Networks}, 33, 41-55.
}

