\name{Bergm-package}
\Rdversion{2.0}
\alias{Bergm-package}
\docType{package}
\title{
Bayesian analysis for exponential random graph models
}
\description{
\code{Bergm} is a collection of functions to fit and diagnose from Bayesian random graph models.
}
\author{
Alberto Caimo <alberto.caimo@ucd.ie> and Nial Friel <nial.friel@ucd.ie>.
}

\references{
Caimo, A. and Friel, N., (2011a) ``Bayesian inference for exponential random graph models,'' Social Networks, 33, 41-55.

Caimo, A. and Friel, N., (2011b) ``Bayesian model selection for exponential random graph models,'' (submitted).

Handcock, M. S., Hunter, D. R., Butts, C. T., Goodreau, S. M., and Morris, M. (2007), ``statnet: Software tools for the representation, visualization, analysis and simulation of network data,'' Journal of Statistical Software, 24,1-11.

Hunter, D. R., Handcock, M. S., Butts, C. T., Goodreau, S. M., and Morris M. (2008), ``ergm: A Package to Fit, Simulate and Diagnose Exponential-Family Models for Networks'', 24, 1-29.
}

