\name{Bergm-package}
\Rdversion{2.1}
\alias{Bergm-package}
\docType{package}
\title{
Bayesian analysis for exponential random graph models
}
\description{
\code{Bergm} is a collection of functions for Bayesian exponential random graph models.
}
\author{
Alberto Caimo <alberto.caimo@ucd.ie> and Nial Friel <nial.friel@ucd.ie>.
}

\references{
Caimo, A. and Friel, N. (2012b), ``Bergm: Bayesian Exponential Random Graphs in R,'' Tech. rep., University College Dublin. Available in e-print format at http://arxiv.org/abs/1201.2770.

Caimo, A. and Friel, N. (2012a), ``Bayesian model selection for exponential random graph models,'' Tech. rep., University College Dublin. Available in e-print format at http://arxiv.org/abs/1201.2337.

Caimo, A. and Friel, N. (2011), ``Bayesian inference for exponential random graph models,'' Social Networks, 33, 41-55.

Hunter, D. R., Handcock, M. S., Butts, C. T., Goodreau, S. M., and Morris M. (2008), ``ergm: A Package to Fit, Simulate and Diagnose Exponential-Family Models for Networks'', Journal of Statistical Software, 24, 1-29.

Handcock, M. S., Hunter, D. R., Butts, C. T., Goodreau, S. M., and Morris, M. (2007), ``statnet: Software tools for the representation, visualization, analysis and simulation of network data,'' Journal of Statistical Software, 24,1-11.
}

