\name{Bergm-package}
\alias{Bergm-package}
\docType{package}
\title{
Bayesian analysis for exponential random graph models
}
\description{
\code{Bergm} is a collection of functions for Bayesian exponential random graph models.
}
\author{
Alberto Caimo <acaimo.stats@gmail.com> and Nial Friel <nial.friel@ucd.ie>.
}

\references{
Caimo, A. and Friel, N. (2013), ``Bergm: Bayesian Exponential Random Graphs in R,'' \emph{Journal of Statistical Software}. Available in e-print format at http://arxiv.org/abs/1201.2770.

Caimo, A. and Friel, N. (2013), ``Bayesian model selection for exponential random graph models,'' \emph{Social Networks}, 35(1), 11-24. Available in e-print format at http://arxiv.org/abs/arXiv:1201.2337.

Caimo, A. and Friel, N. (2011), ``Bayesian inference for exponential random graph models,'' \emph{Social Networks}, 33(1), 41-55. Available in e-print format at http://arxiv.org/abs/arXiv:1007.5192.
}

