\name{Bergm-package}
\Rdversion{1.1}
\alias{Bergm-package}
\docType{package}
\title{
Bayesian inference for exponential random graph models
}
\description{
\code{Bergm} is a collection of functions to fit and diagnose from Bayesian random graph models. For a list of functions type: help(package='Bergm').
}
\author{
Alberto Caimo <alberto.caimo@ucd.ie> and Nial Friel <nial.friel@ucd.ie>
}

\references{
Caimo, A. and Friel, N. (2009) ``Bayesian inference for exponential random graph models,'' (submitted).

Murray, I., Ghahramani, Z., and MacKay, D. (2006), 
``MCMC for doubly-intractable distributions,'' in Proceedings of the 22nd Annual Conference on Uncertainty in Artificial Intelligence (UAI-06), Arlington, Virginia: AUAI Press.

ter Braak, C. J. F., and Vrugt, J. A., ``Differential evolution Markov chain with snooker updater and fewer chains'', Statistics and Computing, 18(4), 435-446, 2008.

Handcock, M. S., Hunter, D. R., Butts, C. T., Goodreau, S. M., and Morris, M. (2007), ``statnet: Software tools for the representation, visualization,
analysis and simulation of network data,'' Journal of Statistical Software, 24,1-11.

Hunter, D. R., Handcock, M. S., Butts, C. T., Goodreau, S. M., and Morris M. (2008), ``ergm: A Package to Fit, Simulate and Diagnose Exponential-Family Models for Networks'', 24, 1-29.
}

