\name{bergm.output}
\alias{bergm.output}
\title{
MCMC output diagnostics for the Bayesian parameter estimation algorithm
}
\description{
This function returns the posterior parameter density estimate and creates simple diagnostic plots for the MCMC produced from a fit. 
}
\usage{
bergm.output(x, 
             ...)
}
\arguments{
  \item{x}{ 
  an \code{R} object of class \code{bergm}.
}
  \item{...}{
additional arguments, to be passed to lower-level functions.
}
}
\seealso{
  \code{\link{bergm}}.
}
