% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bergm.R
\docType{methods}
\name{summary.bergm}
\alias{summary.bergm}
\title{Summary of BERGM posterior output}
\usage{
\method{summary}{bergm}(object, returnTable = FALSE, ...)
}
\arguments{
\item{object}{an \code{R} object of class \code{bergm}.}

\item{returnTable}{logical;
By default FALSE. Should the results table be returned.}

\item{...}{additional arguments, to be passed to lower-level functions.}
}
\description{
This function summarises MCMC output for \code{bergm} objects.
}
